+++ test/authorize/http/response/code.test.js
@@ -16,6 +16,7 @@ describe('http/authorize/response/code', function() {
   it('should create response type without response modes', function(done) {
     var container = new Object();
     container.components = sinon.stub();
+    container.components.withArgs('http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters').returns([]);
     container.components.withArgs('http://i.authnomicon.org/oauth2/authorization/http/ResponseMode').returns([]);
     var acs = new Object();
     acs.issue = sinon.stub().yieldsAsync(null, 'SplxlOBeZQQYbYS6WxSbIA');
@@ -40,6 +41,7 @@ describe('http/authorize/response/code', function() {
   describe('issue', function() {
     var container = new Object();
     container.components = sinon.stub()
+    container.components.withArgs('http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters').returns([]);
     container.components.withArgs('http://i.authnomicon.org/oauth2/authorization/http/ResponseMode').returns([]);
     var acs = new Object();